update test_platform_status_terminate_and_restart_instrument_1_1 to also remove use of ProcessLifecycleEvent and instead ud ResourceAgentLifecycleEvent
